Updating the rendering system. >Starting on looking at "config" files maybe for tiles, items, and settings. > Progress steady
Building up some sort of idea to move on with this framework to.

+++ src/main/java/inferno/saigo/common/tiles/TileData.java
@@ -2,7 +2,7 @@ package inferno.saigo.common.tiles;
 
 public class TileData {
     public Tile tile;
-    public float x, y;
+    public int x, y;
 
 
     @Override
